requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
django-crispy-forms==2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.3
tweepy==4.13.0
tzdata==2023.3
urllib3==1.26.15
